requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
plotly

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud
google-cloud-storage
google-cloud-bigquery
mlflow
s3fs
fastapi
uvicorn

# utilities
six>=1.14
joblib
memoized-property
termcolor

# frontend
streamlit


